Fix build #115: second CompatLib copy + PlayerConnection API mismatch
- Generals/Code/CompatLib/Include/types_compat.h had the identical
  int64/uint64 typedef (`typedef int64_t int64;`) I already fixed in
  GeneralsMD/Code/CompatLib's copy for build #114 -- missed that this
  ZH-build include path pulls in both near-duplicate CompatLib trees.
  Same fix: long long / unsigned long long, matching GNS's steamtypes.h
  exactly.
- PlayerConnection::GetJitter()/ComputeConnectionScore() didn't exist:
  WOLGameSetupMenu.cpp was ported from newer go_client/main and already
  calls both, but NetworkMesh.h/.cpp (deliberately ported from the
  older go_int 32ae5135 snapshot to match already-existing guarded call
  sites elsewhere) predates them. Ported both from go_client/main --
  GetJitter only needs m_vecLatencyHistory (already a member here);
  ComputeConnectionScore only calls the three getters (GetLatency/
  GetJitter/GetConnectionQuality), so it works unmodified against our
  GNS-real-time-status-backed GetConnectionQuality() instead of
  go_client/main's m_vecQualityHistory-averaging one -- no upstream
  m_vecQualityHistory field needed.
- Cross-checked every PlayerConnection/NetworkMesh method actually
  called from WOLGameSetupMenu.cpp/WOLQuickMatchMenu.cpp/the
  GeneralsOnline module against the ported header; everything else
  (GetAllConnections, GetConnectionForUser, getMaximumHistoricalLatency,
  etc.) was already present. The one other newer-API call
  (pMesh->ProcessGameStart) sits behind
  `#if !defined(GENERALS_ONLINE_ENABLE_MATCH_START_COUNTDOWN)`, which
  NextGenMP_defines.h unconditionally defines -- dead code, not compiled.

GeneralsX @bugfix Android port 12/07/2026

+++ Generals/Code/CompatLib/Include/types_compat.h
@@ -97,8 +97,17 @@ typedef float FLOAT;
 // Removed duplicate definitions to avoid conflicts
 // typedef int64_t _int64;   // REMOVED - use bittype.h
 // typedef uint64_t _uint64; // REMOVED - use bittype.h
-typedef int64_t int64;
-typedef uint64_t uint64;
+// GeneralsX @bugfix Android port 12/07/2026 - long long (not int64_t) to match
+// Valve GameNetworkingSockets' own steamtypes.h non-Windows typedefs exactly.
+// On LP64 platforms (Android/Linux arm64) int64_t is `long`, but steamtypes.h
+// unconditionally does `typedef long long int64;` with no #ifndef guard --
+// different underlying type, same name, hard redefinition error. long long
+// is 64-bit everywhere we target, so this is a spelling change only. (Mirrors
+// the identical fix in GeneralsMD/Code/CompatLib/Include/types_compat.h --
+// this file is the ZH build's *other* CompatLib on the include path, kept in
+// sync historically as a near-duplicate.)
+typedef long long int64;
+typedef unsigned long long uint64;
 
 typedef int32_t *LPARAM;
 typedef size_t WPARAM;
